+++ src/test/java/racingcar/model/TrialNumTest.java
@@ -1,7 +1,6 @@
 package racingcar.model;
 
 import org.junit.jupiter.api.Test;
-import racingcar.model.vo.TrialNum;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
